+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: akisuzuk <akisuzuk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/14 09:35:53 by akisuzuk          #+#    #+#             */
-/*   Updated: 2023/03/25 13:14:36 by akisuzuk         ###   ########.fr       */
+/*   Updated: 2023/03/25 16:12:07 by akisuzuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,8 +39,8 @@ void	ft_putnbr_fd(int n, int fd);
 void	ft_putnbr_base_long(long nbr, char *base);
 
 void	ft_putchar_rep(char c, int fd, int rep);
-int		ft_get_digit(int num, int base);
-void	ft_get_putlen(t_flag *info, int num);
+int		ft_get_digit(long num, int base);
+void	ft_get_putlen(t_flag *info, long num);
 
 void	ft_print_percent(t_flag *info);
 void	ft_unsigned_print(va_list *arg, t_flag *info);
